telegram trader
With the implementation of the new `Trader` class we are sending all the orders to the server at once (in parallel) instead of doing it the old way which was sending them one at a time (in series). The old way took about ten (10) seconds to create sixteen (16) orders and the new way does it about five (5) times faster so it now only takes about two (2) seconds.